Move the bulk_invoke functions into their own headers
Fixes #19
